Simplify GraphicsFuzz and allow compilation with Java 17 (#1)
This change removes lots of legacy infrastructure from the
GraphicsFuzz project - everything to do with workers and servers. The
idea is that gfauto should be the route for using GraphicsFuzz (which
currently means that only testing of Vulkan drivers is supported).

As part of this change, the Java compiler has changed to Java 17. This
will allow Kotlin to be used in the project in the future.